pyproject.toml
[project]
name = "comfyui-sam3"
description = "SAM 3 (Segment Anything Model 3) integration for ComfyUI - advanced image and video segmentation with text prompts, boxes, and points. This node is a first draft, please open an issue on GitHub if you encounter any :)"
version = "0.1.0"
license = { file = "LICENSE" }
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=1.0.17",
    "ftfy==6.1.1",
    "regex",
    "iopath>=0.1.10",
    "huggingface_hub",
    "opencv-python>=4.8.0",
    "psutil>=5.9.0",
    "pycocotools>=2.0.6",
    "scikit-image>=0.19.0",
    "numpy>=1.26",
    "Pillow",
    "tqdm",
    "typing_extensions"
]

[project.urls]
Repository = "https://github.com/PozzettiAndrea/ComfyUI-SAM3"
Documentation = "https://github.com/PozzettiAndrea/ComfyUI-SAM3/blob/main/README.md"

[tool.comfy]
PublisherId = "pznodes"
DisplayName = "ComfyUI-SAM3"
Icon = ""

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
